# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/mentor-requests": {
      "get": {
        "operationId": "MentorRequestsController_getAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter mentor requests by project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mentorId",
            "required": false,
            "in": "query",
            "description": "Filter mentor requests by mentor ID or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter mentor requests by status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of mentor requests returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MentorRequestEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "summary": "List mentor requests",
        "tags": [
          "Mentor Requests"
        ]
      },
      "post": {
        "operationId": "MentorRequestsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMentorRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mentor request created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "summary": "Create a new mentor request",
        "tags": [
          "Mentor Requests"
        ]
      }
    },
    "/mentor-requests/{id}": {
      "get": {
        "operationId": "MentorRequestsController_getOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mentor request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor request returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Mentor request not found."
          }
        },
        "summary": "Get a mentor request by ID",
        "tags": [
          "Mentor Requests"
        ]
      },
      "put": {
        "operationId": "MentorRequestsController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mentor request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMentorRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mentor request updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Mentor request not found."
          }
        },
        "summary": "Update a mentor request",
        "tags": [
          "Mentor Requests"
        ]
      },
      "delete": {
        "operationId": "MentorRequestsController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mentor request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor request deleted successfully."
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Mentor request not found."
          }
        },
        "summary": "Delete a mentor request (Admin only)",
        "tags": [
          "Mentor Requests"
        ]
      }
    },
    "/mentor-requests/{id}/accept": {
      "put": {
        "operationId": "MentorRequestsController_accept",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mentor request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor request accepted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Mentor request not found."
          }
        },
        "summary": "Accept a mentor request",
        "tags": [
          "Mentor Requests"
        ]
      }
    },
    "/mentor-requests/{id}/decline": {
      "put": {
        "operationId": "MentorRequestsController_decline",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mentor",
                "project-owner",
                "collaborator",
                "superuser"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mentor request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mentor request declined successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Mentor request not found."
          }
        },
        "summary": "Decline a mentor request",
        "tags": [
          "Mentor Requests"
        ]
      }
    },
...
    "/mentor-marketplace/mentors/{id}": {
      "get": {
        "operationId": "MentorMarketplaceController_getMentor",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mentor profile detail + reviews",
        "tags": [
          "Mentor Marketplace"
        ]
      }
    },
...
  },
  "info": {
    "title": "TeamForge API",
    "description": "The Student Project Collaboration Platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MentorRequestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique mentor request identifier",
            "example": "4f3a9e2d-d9e8-4bbb-b6f8-a3b132b8f1c1"
          },
          "projectId": {
            "type": "string",
            "description": "Target project ID for this mentor request",
            "example": "c9b1d92a-a73f-4a4c-8b0f-2d2f78a8de12"
          },
          "mentorId": {
            "type": "string",
            "description": "Target mentor user ID",
            "example": "a1d2c3b4-e5f6-7890-1234-56789abcdef0"
          },
          "mentorEmail": {
            "type": "string",
            "description": "Target mentor email address",
            "example": "mentor@example.com"
          },
          "message": {
            "type": "string",
            "description": "Optional project owner message or pitch for the mentor",
            "example": "We would love your guidance on our full-stack project with React and Node."
          },
          "status": {
            "type": "string",
            "description": "Current mentor request status",
            "enum": [
              "pending",
              "accepted",
              "declined"
            ],
            "example": "pending"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp when the request was created",
            "example": "2026-04-23T12:34:56.789Z"
          },
          "acceptedAt": {
            "type": "string",
            "description": "ISO timestamp when the request was accepted",
            "example": "2026-04-24T08:15:00.123Z"
          },
          "declinedAt": {
            "type": "string",
            "description": "ISO timestamp when the request was declined",
            "example": "2026-04-24T09:00:00.123Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "status",
          "createdAt"
        ]
      },
      "CreateMentorRequestDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Target project ID for the mentor request",
            "example": "c9b1d92a-a73f-4a4c-8b0f-2d2f78a8de12"
          },
          "mentorId": {
            "type": "string",
            "description": "Target mentor user ID",
            "example": "a1d2c3b4-e5f6-7890-1234-56789abcdef0"
          },
          "mentorEmail": {
            "type": "string",
            "description": "Target mentor email address",
            "example": "mentor@example.com"
          },
          "message": {
            "type": "string",
            "description": "Optional message or pitch shared with the mentor",
            "example": "We would love your guidance on our full-stack project with React and Node."
          }
        },
        "required": [
          "projectId"
        ]
      },
      "UpdateMentorRequestDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Target project ID for the mentor request",
            "example": "c9b1d92a-a73f-4a4c-8b0f-2d2f78a8de12"
          },
          "mentorId": {
            "type": "string",
            "description": "Target mentor user ID",
            "example": "a1d2c3b4-e5f6-7890-1234-56789abcdef0"
          },
          "mentorEmail": {
            "type": "string",
            "description": "Target mentor email address",
            "example": "mentor@example.com"
          },
          "message": {
            "type": "string",
            "description": "Optional message or pitch shared with the mentor",
            "example": "We would love your guidance on our full-stack project with React and Node."
          },
          "status": {
            "type": "string",
            "description": "Mentor request status",
            "enum": [
              "pending",
              "accepted",
              "declined"
            ],
            "example": "accepted"
          }
        }
      },
...
    }
  }
}